test/fixtures/wpt/performance-timeline/not-clonable.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  const t = async_test("Test that a postMessage of a performance entry fails");
  addEventListener("message", t.step_func_done(e => {
    assert_equals(e.data, "PASS");
  }));
</script>
<iframe src="resources/postmessage-entry.html"></iframe>
